# integrations/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rahoitusrekisteri",
    "description": "Tampereen ulkoisen rahoituksen tietokanta. Lähteiden ilmoitukset ja asiantuntijan arviot säilytetään erillisinä. API v1 käyttää pysyviä tunnisteita ja sivutusta.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/sources": {
      "get": {
        "tags": [
          "Lähteet"
        ],
        "summary": "Sources",
        "operationId": "sources_v1_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "EntraID": []
          }
        ]
      }
    },
...
    "/v1/runs": {
      "get": {
        "tags": [
          "Tila"
        ],
        "summary": "Runs",
        "operationId": "runs_v1_runs_get",
        "security": [
          {
            "EntraID": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/facets": {
      "get": {
        "tags": [
          "Haut"
        ],
        "summary": "Facets",
        "operationId": "facets_v1_facets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "EntraID": []
          }
        ]
      }
    },
    "/v1/funders": {
      "get": {
        "tags": [
          "Lähteet"
        ],
        "summary": "Funders",
        "operationId": "funders_v1_funders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "EntraID": []
          }
        ]
      }
    },
    "/v1/priorities": {
      "get": {
        "tags": [
          "Lähteet"
        ],
        "summary": "Priorities",
        "operationId": "priorities_v1_priorities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "EntraID": []
          }
        ]
      }
    },
...
    "/v1/opportunities/{oid}": {
      "get": {
        "tags": [
          "Haut"
        ],
        "summary": "Opportunity",
        "operationId": "opportunity_v1_opportunities__oid__get",
        "security": [
          {
            "EntraID": []
          }
        ],
        "parameters": [
          {
            "name": "oid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Oid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/opportunities/{oid}/evidence": {
      "get": {
        "tags": [
          "Haut"
        ],
        "summary": "Evidence",
        "operationId": "evidence_v1_opportunities__oid__evidence_get",
        "security": [
          {
            "EntraID": []
          }
        ],
        "parameters": [
          {
            "name": "oid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Oid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/opportunities/{oid}/documents": {
      "get": {
        "tags": [
          "Ehdot ja liitteet"
        ],
        "summary": "Documents",
        "operationId": "documents_v1_opportunities__oid__documents_get",
        "security": [
          {
            "EntraID": []
          }
        ],
        "parameters": [
          {
            "name": "oid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Oid"
            }
          },
          {
            "name": "include_history",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include History"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{did}/text": {
      "get": {
        "tags": [
          "Ehdot ja liitteet"
        ],
        "summary": "Document Text",
        "operationId": "document_text_v1_documents__did__text_get",
        "security": [
          {
            "EntraID": []
          }
        ],
        "parameters": [
          {
            "name": "did",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Did"
            }
          },
          {
            "name": "sha256",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[a-f0-9]{64}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sha256"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{did}/original": {
      "get": {
        "tags": [
          "Ehdot ja liitteet"
        ],
        "summary": "Document Original",
        "operationId": "document_original_v1_documents__did__original_get",
        "security": [
          {
            "EntraID": []
          }
        ],
        "parameters": [
          {
            "name": "did",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Did"
            }
          },
          {
            "name": "sha256",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[a-f0-9]{64}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sha256"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/changes": {
      "get": {
        "tags": [
          "Muutokset"
        ],
        "summary": "Changes",
        "operationId": "changes_v1_changes_get",
        "security": [
          {
            "EntraID": []
          }
        ],
        "parameters": [
          {
            "name": "after_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After Id"
            }
          },
          {
            "name": "through_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Through Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "opportunity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opportunity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/snapshots/{snapshot_id}": {
      "get": {
        "tags": [
          "Lähteet"
        ],
        "summary": "Snapshot",
        "operationId": "snapshot_v1_snapshots__snapshot_id__get",
        "security": [
          {
            "EntraID": []
          }
        ],
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/opportunities/{oid}/review": {
      "put": {
        "tags": [
          "Asiantuntijan arvio"
        ],
        "summary": "Review",
        "operationId": "review_v1_opportunities__oid__review_put",
        "security": [
          {
            "EntraID": []
          }
        ],
        "parameters": [
          {
            "name": "oid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Oid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ReviewInput": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version",
            "description": "Current review version; zero for a new review"
          },
          "source_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Source Version",
            "description": "Opportunity version the reviewer actually read"
          },
          "eligibility": {
            "type": "string",
            "enum": [
              "unreviewed",
              "eligible",
              "conditional",
              "ineligible"
            ],
            "title": "Eligibility"
          },
          "city_role": {
            "type": "string",
            "enum": [
              "unknown",
              "lead",
              "partner",
              "beneficiary"
            ],
            "title": "City Role",
            "default": "unknown"
          },
          "service_area": {
            "type": "string",
            "maxLength": 200,
            "title": "Service Area",
            "default": ""
          },
          "notes": {
            "type": "string",
            "maxLength": 20000,
            "title": "Notes",
            "default": ""
          },
          "evidence_url": {
            "type": "string",
            "maxLength": 2048,
            "title": "Evidence Url",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "expected_version",
          "source_version",
          "eligibility"
        ],
        "title": "ReviewInput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "EntraID": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}